Move notifications handling logic to notificationsApi, where onCacheEntryAdded fn listens to notifications events and updates cache.
