+++ src/features/notifications/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from './useNotificationsController'
-export * from './useNotificationsSocket'
